Revert "feat: routing"
This reverts commit 14a2abf8e0844d1b87a565ff9dcd9e07ac963b90.
